优化代码结构

+++ flapypan-blog-api/src/main/kotlin/top/flapypan/blog/service/TagService.kt
@@ -1,6 +1,5 @@
 package top.flapypan.blog.service
 
-import org.springframework.data.domain.Example
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
@@ -8,12 +7,12 @@ import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import top.flapypan.blog.common.RestException
-import top.flapypan.blog.controller.TagAddRequest
-import top.flapypan.blog.controller.TagUpdateRequest
-import top.flapypan.blog.dto.ArticleSimpleDTO
 import top.flapypan.blog.entity.Tag
 import top.flapypan.blog.repository.ArticleRepository
 import top.flapypan.blog.repository.TagRepository
+import top.flapypan.blog.vo.ArticleInfo
+import top.flapypan.blog.vo.TagAddRequest
+import top.flapypan.blog.vo.TagUpdateRequest
 
 @Service
 class TagService(
@@ -21,38 +20,40 @@ class TagService(
     private val articleRepository: ArticleRepository
 ) {
 
-    fun getAll() = repository.getTagDTOList()
+    fun getAll(): List<Tag> = repository.findAll()
 
     fun findByName(name: String) =
         repository.findByName(name) ?: throw RestException(HttpStatus.NOT_FOUND.value(), "不存在的标签")
 
     @Transactional
-    fun add(addRequest: TagAddRequest): Tag {
+    fun add(addRequest: TagAddRequest): String {
         val tag = addRequest.createEntity()
-        if (repository.exists(Example.of(tag))) {
+        if (repository.existsByNameIgnoreCase(addRequest.name)) {
             throw RestException(HttpStatus.BAD_REQUEST.value(), "标签已经存在")
         }
-        return repository.save(tag)
+        return repository.save(tag).name
     }
 
     @Transactional
-    fun update(updateRequest: TagUpdateRequest): Tag {
+    fun update(updateRequest: TagUpdateRequest): String {
         val tag = repository.findByIdOrNull(updateRequest.id)
             ?: throw RestException(HttpStatus.NOT_FOUND.value(), "不存在的标签")
         updateRequest.copyToEntity(tag)
-        return repository.save(tag)
+        return repository.save(tag).name
     }
 
     @Transactional
     fun delete(id: Long) {
-        repository.deleteArticleTagByTagId(id)
+        if (articleRepository.countByTagsId(id) > 0) {
+            throw RestException(HttpStatus.BAD_REQUEST.value(), "当前标签下存在文章")
+        }
         repository.deleteById(id)
     }
 
-    fun getArticleByTag(tagName: String, pageable: Pageable): Page<ArticleSimpleDTO> {
+    fun getArticleByTag(tagName: String, pageable: Pageable): Page<ArticleInfo> {
         repository.findByName(tagName) ?: RestException(HttpStatus.NOT_FOUND.value(), "不存在的标签")
         return articleRepository
             .findAllByTagsName(tagName, pageable)
-            .map(::ArticleSimpleDTO)
+            .map(::ArticleInfo)
     }
 }
